Add module logic and sample methods
