+++ lib/modules/calculateable.rb
@@ -1,0 +1,34 @@
+module Calculateable
+  def average_goals(goals_hash)
+    average_goals = {}
+    goals_hash.each do |team, tot_score|
+      average_goals[team] = (tot_score.to_f / games_by_team[team]).round(2)
+    end
+
+    average_goals
+  end
+
+  def games_by_team
+    @game_collection.collection.inject(Hash.new(0)) do |count, game|
+      count[game[1].home_team_id] += 1
+      count[game[1].away_team_id] += 1
+      count
+    end
+  end
+
+  def goals_by_team
+    @game_collection.collection.inject(Hash.new(0)) do |scores, game|
+      scores[game[1].home_team_id] += game[1].home_goals.to_i
+      scores[game[1].away_team_id] += game[1].away_goals.to_i
+      scores
+    end
+  end
+
+  def goals_against_team
+    @game_collection.collection.inject(Hash.new(0)) do |scores, game|
+      scores[game[1].home_team_id] += game[1].away_goals.to_i
+      scores[game[1].away_team_id] += game[1].home_goals.to_i
+      scores
+    end
+  end
+end
